Util: Add a retry method to handle intermittent errors
`util.retry()` will invoke a promise-returning method and if the promise is
rejected, it will try again in 2 seconds. If that promise is rejected, it will
try again in 4 seconds. If that promise is rejected, the error will propagate.

This should hopefully fix the error where the PR branch doesn't exist by the
time we process the webhook.

Fixes gh-23