All: Improve name validation
Replaces somewhat hostile language for name validation by pointing at
the required process.

Also removes the commit name+email validation, since that ignores the
manual verification and is redundant anyway. It was supposed to speed
up fixing issues with git author and signature at the time, but that's
acceptable to loose.

Ref #54